refactor: products앱 이름 product로 수정
- domain도 products -> product로 수정
- mysql내부의 장고관련 테이블 속성들도 수정
