+++ product/admin.py
