+++ product/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class ProductsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'products'
+    name = 'product'
